Does file use a d2 extension?

is_d2_file() tests for a ".d2" file extension. A d2 file extension is not required in all cases but most functions in this package expect a d2 extension by default.

Usage

is_d2_file(file, fileext = "d2")

Arguments

file

File name(s) to test.

fileext

File extension(s) to test for.